initial commit of local branch
Created initial preview of website.
Project guidlines state that the project will need to be pure HTML and CSS with no JavaScript.

Started styling using CSS but will need to be updated to Scss.

Created initial loading animation using keyframes and a final width of 0% to effectivally remove the div without the use of JS.

Pending tasks:
The header is not displaying its box-border, check why/if this is possible.
The heart/favorite icon is not positioned correctly, particularly in regards to responsiveness. Impliment grid or flexbox to remedy.
Animations for hover need to be made.

menu pages for each restaurant need to be made.