Fixed heart hover effect (now fills on hover)
Added a fill effect to the hearts on hover (on restaurant cards).
Used fontawesome icon of full heart with color: transparent , then added a webkit stroke to outline the heart in grey to match the mock. Added hover color change with animation delay to resemble the mock.

Adjusted margins of how-it-works container