Changed font sizes, weights, adjusted padding, general mock mathching using new strategy
With differences highlighted between my version and the mock I utilized a combination of the PixelPerfect extension and Gimp to ensure matching font sizes and weights, as well as matching padding.

Since no font sizes were provided I brought the mock into Gimp and installed all associated fonts. I was then able to create an overlapping text and match the sizing and weight perfectly. This was done for all font on the index page.

Margins and padding were all adjusted.

Restaurant NEW box adjusted in size, position, and text color.